Adding serialization support for ExecutableSequence class.
