+++ src/test/java/randoop/test/SequenceTests.java
@@ -8,9 +8,7 @@ import java.io.*;
 import java.util.*;
 
 import randoop.DummyVisitor;
-import randoop.ExecutionOutcome;
 import randoop.Globals;
-import randoop.NormalExecution;
 import randoop.contract.EqualsHashcode;
 import randoop.contract.EqualsReflexive;
 import randoop.contract.EqualsSymmetric;
@@ -228,142 +226,5 @@ public class SequenceTests {
       }
     }
     return trimmed;
-  }
-
-  /**
-   * Tests the serialization mechanisms of Randoop by serializing and de-serializing an Executable Sequence
-   *
-   */
-  @Test
-  public void testSerialization()
-      throws NoSuchMethodException, IOException, ClassNotFoundException {
-    final int SIZE_STATEMENT_INDEX = 3;
-    ExecutableSequence es = createExecutableSequenceForSerialization();
-    es.execute(new DummyVisitor(), new DummyCheckGenerator());
-    ExecutionOutcome result = es.getResult(SIZE_STATEMENT_INDEX);
-
-    if (result instanceof NormalExecution) {
-      Object valueBeforeSerialization = ((NormalExecution) result).getRuntimeValue();
-      String codeBeforeSerialization = es.toCodeString();
-
-      ByteArrayOutputStream out = getSerializedExecutableSequence(es);
-      ByteArrayInputStream in = new ByteArrayInputStream(out.toByteArray());
-
-      ExecutableSequence esDeserialized = getDeserializedExecutableSequence(in);
-
-      if (esDeserialized == null) {
-        fail("Deserialezed to null sequence.");
-      }
-
-      esDeserialized.execute(new DummyVisitor(), new DummyCheckGenerator());
-      ExecutionOutcome result2 = esDeserialized.getResult(SIZE_STATEMENT_INDEX);
-
-      if (result2 instanceof NormalExecution) {
-        Object valueAfterSerialization = ((NormalExecution) result2).getRuntimeValue();
-        String codeAfterSerialization = es.toCodeString();
-
-        assertEquals(valueBeforeSerialization, valueAfterSerialization);
-        assertEquals(codeBeforeSerialization, codeAfterSerialization);
-      } else {
-        fail("Deserialized sequence not resulting in normal execution.");
-      }
-
-    } else {
-      fail("Sequence not resulting in normal execution.");
-    }
-  }
-
-  /***
-   * Produces the following sequence found on Randoop Developer's manual :
-   *  LinkedList<String> l = new LinkedList<>();
-   * String str = "hi!";
-   * l.addFirst(str);
-   * int i = l.size();
-   */
-  private static Sequence createSequence() throws NoSuchMethodException {
-    Sequence s = new Sequence();
-
-    InstantiatedType linkedListType = JDKTypes.LINKED_LIST_TYPE.instantiate(JavaTypes.STRING_TYPE);
-    Substitution<ReferenceType> substLL = linkedListType.getTypeSubstitution();
-    TypedOperation newLL =
-        TypedOperation.forConstructor(LinkedList.class.getConstructor()).apply(substLL);
-
-    TypedOperation newLiteral =
-        TypedOperation.createPrimitiveInitialization(JavaTypes.STRING_TYPE, "hi!");
-    TypedOperation addFirst =
-        TypedOperation.forMethod(LinkedList.class.getMethod("addFirst", Object.class))
-            .apply(substLL);
-    TypedOperation size =
-        TypedOperation.forMethod(LinkedList.class.getMethod("size")).apply(substLL);
-
-    InstantiatedType treeSetType = JDKTypes.TREE_SET_TYPE.instantiate(JavaTypes.STRING_TYPE);
-    Substitution<ReferenceType> substTS = treeSetType.getTypeSubstitution();
-    TypedOperation wcTS =
-        TypedOperation.forConstructor(TreeSet.class.getConstructor(Collection.class))
-            .apply(substTS)
-            .applyCaptureConversion();
-    Substitution<ReferenceType> substWC =
-        Substitution.forArgs(wcTS.getTypeParameters(), (ReferenceType) JavaTypes.STRING_TYPE);
-    TypedOperation newTS = wcTS.apply(substWC);
-    TypedOperation syncA =
-        TypedOperation.forMethod(Collections.class.getMethod("synchronizedSet", Set.class));
-    Substitution<ReferenceType> substA =
-        Substitution.forArgs(syncA.getTypeParameters(), (ReferenceType) JavaTypes.STRING_TYPE);
-    TypedOperation syncS = syncA.apply(substA);
-
-    s = s.extend(newLL);
-    s = s.extend(newLiteral);
-    s = s.extend(addFirst, s.getVariable(0), s.getVariable(1));
-    s = s.extend(size, s.getVariable(0));
-    s = s.extend(newTS, s.getVariable(0));
-    s = s.extend(syncS, s.getVariable(4));
-
-    return s;
-  }
-
-  private static ExecutableSequence createExecutableSequenceForSerialization()
-      throws NoSuchMethodException {
-    Sequence sampleSeq = createSequence();
-
-    ExecutableSequence es = new ExecutableSequence(sampleSeq);
-
-    return es;
-  }
-
-  public static ByteArrayOutputStream getSerializedExecutableSequence(ExecutableSequence es)
-      throws IOException {
-    ByteArrayOutputStream bos = new ByteArrayOutputStream();
-
-    ObjectOutput out;
-    try {
-      out = new ObjectOutputStream(bos);
-      out.writeObject(es);
-      out.flush();
-    } finally {
-      bos.close();
-    }
-
-    return bos;
-  }
-
-  public static ExecutableSequence getDeserializedExecutableSequence(ByteArrayInputStream input)
-      throws IOException, ClassNotFoundException {
-    ExecutableSequence es = null;
-
-    ObjectInputStream ois = null;
-    try {
-      ois = new ObjectInputStream(input);
-      Object des = ois.readObject();
-      es = (ExecutableSequence) des;
-    } finally {
-      if (input != null) {
-        input.close();
-      }
-      if (ois != null) {
-        ois.close();
-      }
-    }
-
-    return es;
   }
 }
